Move SQLite shared code to db
