+++ collect_app/src/main/java/org/odk/collect/android/database/entities/DatabaseEntitiesRepository.kt
@@ -7,18 +7,18 @@ import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteException
 import android.provider.BaseColumns._ID
 import androidx.core.database.sqlite.transaction
-import org.odk.collect.androidshared.sqlite.CursorExt.first
-import org.odk.collect.androidshared.sqlite.CursorExt.foldAndClose
-import org.odk.collect.androidshared.sqlite.CursorExt.getInt
-import org.odk.collect.androidshared.sqlite.CursorExt.getIntOrNull
-import org.odk.collect.androidshared.sqlite.CursorExt.getString
-import org.odk.collect.androidshared.sqlite.CursorExt.getStringOrNull
-import org.odk.collect.androidshared.sqlite.DatabaseConnection
-import org.odk.collect.androidshared.sqlite.DatabaseMigrator
-import org.odk.collect.androidshared.sqlite.SQLiteColumns.ROW_ID
-import org.odk.collect.androidshared.sqlite.SQLiteDatabaseExt.delete
-import org.odk.collect.androidshared.sqlite.SQLiteDatabaseExt.query
-import org.odk.collect.androidshared.sqlite.SQLiteUtils
+import org.odk.collect.db.sqlite.CursorExt.first
+import org.odk.collect.db.sqlite.CursorExt.foldAndClose
+import org.odk.collect.db.sqlite.CursorExt.getInt
+import org.odk.collect.db.sqlite.CursorExt.getIntOrNull
+import org.odk.collect.db.sqlite.CursorExt.getString
+import org.odk.collect.db.sqlite.CursorExt.getStringOrNull
+import org.odk.collect.db.sqlite.DatabaseConnection
+import org.odk.collect.db.sqlite.DatabaseMigrator
+import org.odk.collect.db.sqlite.SQLiteColumns.ROW_ID
+import org.odk.collect.db.sqlite.SQLiteDatabaseExt.delete
+import org.odk.collect.db.sqlite.SQLiteDatabaseExt.query
+import org.odk.collect.db.sqlite.SQLiteUtils
 import org.odk.collect.entities.storage.EntitiesRepository
 import org.odk.collect.entities.storage.Entity
