+++ collect_app/src/main/java/org/odk/collect/android/database/forms/FormDatabaseMigrator.java
@@ -3,8 +3,8 @@ package org.odk.collect.android.database.forms;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 
-import org.odk.collect.androidshared.sqlite.DatabaseMigrator;
-import org.odk.collect.androidshared.sqlite.SQLiteUtils;
+import org.odk.collect.db.sqlite.DatabaseMigrator;
+import org.odk.collect.db.sqlite.SQLiteUtils;
 
 import static android.provider.BaseColumns._ID;
 import static org.odk.collect.android.database.DatabaseConstants.FORMS_TABLE_NAME;
