+++ collect_app/src/main/java/org/odk/collect/android/database/savepoints/DatabaseSavepointsRepository.kt
@@ -8,10 +8,10 @@ import org.odk.collect.android.database.DatabaseConstants.SAVEPOINTS_DATABASE_VE
 import org.odk.collect.android.database.DatabaseConstants.SAVEPOINTS_TABLE_NAME
 import org.odk.collect.android.database.savepoints.DatabaseSavepointsColumns.FORM_DB_ID
 import org.odk.collect.android.database.savepoints.DatabaseSavepointsColumns.INSTANCE_DB_ID
-import org.odk.collect.androidshared.sqlite.CursorExt.foldAndClose
-import org.odk.collect.androidshared.sqlite.DatabaseConnection
-import org.odk.collect.androidshared.sqlite.SQLiteDatabaseExt.delete
-import org.odk.collect.androidshared.sqlite.SQLiteDatabaseExt.query
+import org.odk.collect.db.sqlite.CursorExt.foldAndClose
+import org.odk.collect.db.sqlite.DatabaseConnection
+import org.odk.collect.db.sqlite.SQLiteDatabaseExt.delete
+import org.odk.collect.db.sqlite.SQLiteDatabaseExt.query
 import org.odk.collect.forms.savepoints.Savepoint
 import org.odk.collect.forms.savepoints.SavepointsRepository
 import org.odk.collect.shared.PathUtils
