+++ collect_app/src/main/java/org/odk/collect/android/dynamicpreload/ExternalSQLiteOpenHelper.java
@@ -31,11 +31,11 @@ import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
 
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.androidshared.sqlite.AltDatabasePathContext;
+import org.odk.collect.db.sqlite.AltDatabasePathContext;
 import org.odk.collect.android.exception.ExternalDataException;
-import org.odk.collect.androidshared.sqlite.CustomSQLiteQueryBuilder;
-import org.odk.collect.androidshared.sqlite.CustomSQLiteQueryExecutor;
-import org.odk.collect.androidshared.sqlite.SQLiteUtils;
+import org.odk.collect.db.sqlite.CustomSQLiteQueryBuilder;
+import org.odk.collect.db.sqlite.CustomSQLiteQueryExecutor;
+import org.odk.collect.db.sqlite.SQLiteUtils;
 import org.odk.collect.shared.strings.Md5;
 
 import java.io.File;
