+++ collect_app/src/main/java/org/odk/collect/android/fastexternalitemset/ItemsetDbAdapter.java
@@ -7,7 +7,7 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.androidshared.sqlite.AltDatabasePathContext;
+import org.odk.collect.db.sqlite.AltDatabasePathContext;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
 import org.odk.collect.shared.PathUtils;
