+++ collect_app/src/test/java/org/odk/collect/android/application/RobolectricApplication.java
@@ -9,7 +9,7 @@ import androidx.test.core.app.ApplicationProvider;
 import androidx.work.Configuration;
 import androidx.work.WorkManager;
 
-import org.odk.collect.androidshared.sqlite.DatabaseConnection;
+import org.odk.collect.db.sqlite.DatabaseConnection;
 import org.odk.collect.androidshared.ui.multiclicksafe.MultiClickGuard;
 import org.odk.collect.crashhandler.CrashHandler;
 import org.robolectric.shadows.ShadowApplication;
