+++ collect_app/src/test/java/org/odk/collect/android/database/FormDatabaseMigratorTest.java
@@ -11,7 +11,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.database.forms.FormDatabaseMigrator;
-import org.odk.collect.androidshared.sqlite.SQLiteUtils;
+import org.odk.collect.db.sqlite.SQLiteUtils;
 
 import java.util.List;
