+++ db/build.gradle.kts
@@ -41,5 +41,13 @@ android {
 
 dependencies {
     coreLibraryDesugaring(Dependencies.desugar)
+
+    implementation(project(":shared"))
     implementation(Dependencies.kotlin_stdlib)
+    implementation(Dependencies.androidx_core_ktx)
+    implementation(Dependencies.timber)
+
+    testImplementation(Dependencies.junit)
+    testImplementation(Dependencies.androidx_test_ext_junit)
+    testImplementation(Dependencies.robolectric)
 }
