+++ db/src/main/java/org/odk/collect/db/sqlite/CustomSQLiteQueryExecutor.java
@@ -1,4 +1,4 @@
-package org.odk.collect.androidshared.sqlite;
+package org.odk.collect.db.sqlite;
 
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
