+++ db/src/main/java/org/odk/collect/db/sqlite/DatabaseConnection.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.androidshared.sqlite
+package org.odk.collect.db.sqlite
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
