+++ db/src/main/java/org/odk/collect/db/sqlite/SQLiteColumns.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.androidshared.sqlite
+package org.odk.collect.db.sqlite
 
 object SQLiteColumns {
     const val ROW_ID = "rowid"
