+++ db/src/test/java/org/odk/collect/db/sqlite/CustomSQLiteQueryBuilderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.odk.collect.androidshared.sqlite;
+package org.odk.collect.db.sqlite;
 
 import static org.junit.Assert.assertEquals;
