+++ db/src/test/java/org/odk/collect/db/sqlite/SQLiteUtilsTest.java
@@ -1,13 +1,14 @@
-package org.odk.collect.androidshared.sqlite;
+package org.odk.collect.db.sqlite;
 
 import static android.content.Context.MODE_PRIVATE;
-import static junit.framework.TestCase.assertFalse;
 import static junit.framework.TestCase.assertTrue;
 
 import android.database.sqlite.SQLiteDatabase;
 
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+
+import junit.framework.TestCase;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -28,7 +29,7 @@ public class SQLiteUtilsTest {
         assertTrue(SQLiteUtils.doesColumnExist(db, tableName, "id"));
         assertTrue(SQLiteUtils.doesColumnExist(db, tableName, "column1"));
 
-        assertFalse(SQLiteUtils.doesColumnExist(db, tableName, "column2"));
+        TestCase.assertFalse(SQLiteUtils.doesColumnExist(db, tableName, "column2"));
     }
 
     @Test
@@ -38,7 +39,7 @@ public class SQLiteUtilsTest {
 
         SQLiteDatabase db = ApplicationProvider.getApplicationContext().openOrCreateDatabase("testDatabase", MODE_PRIVATE, null);
 
-        assertFalse(SQLiteUtils.doesTableExist(db, tableName));
+        TestCase.assertFalse(SQLiteUtils.doesTableExist(db, tableName));
 
         // Create the table and check again
         List<String> columnDefinitions = new ArrayList<>();
